pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "mysql-connector-python>=9.3.0",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "pykrx>=1.0.51",
    "yfinance>=0.2.65",
]
